Fixed some issues for chat
