+++ src/screens/ChatRoom/ChatRoom.tsx
@@ -11,19 +11,12 @@ import React, {
 } from 'react';
 import {
   View,
-  Image,
   LogBox,
-  TouchableOpacity,
-  TextInput,
-  Platform,
-  KeyboardAvoidingView,
   FlatList,
   Keyboard,
-  ActivityIndicator,
   LayoutAnimation,
 } from 'react-native';
 import ImageView from 'react-native-image-viewing';
-import CustomText from '../../components/CustomText';
 import { useStyles } from './styles';
 import {
   type RouteProp,
@@ -34,7 +27,6 @@ import {
 } from '@react-navigation/native';
 import type { RootStackParamList } from '../../routes/RouteParamList';
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import BackButton from '../../components/BackButton';
 import {
   MessageContentType,
   MessageRepository,
@@ -50,26 +42,14 @@ import ImagePicker, {
   type Asset,
   launchCamera,
 } from 'react-native-image-picker';
-// import { SafeAreaView } from 'react-native-safe-area-context';
 import LoadingImage from '../../components/LoadingImage';
 import EditMessageModal from '../../components/EditMessageModal';
-// import { GroupChatIcon } from '../../svg/GroupChatIcon';
-import { AvatarIcon } from '../../svg/AvatarIcon';
-import { CameraBoldIcon } from '../../svg/CameraBoldIcon';
-// import { MenuIcon } from '../../svg/MenuIcon';
-// import { PlusIcon } from '../../svg/PlusIcon';
-import { SendChatIcon } from '../../svg/SendChatIcon';
-import { AlbumIcon } from '../../svg/AlbumIcon';
-import { useTheme } from 'react-native-paper';
-import type { MyMD3Theme } from '../../providers/amity-ui-kit-provider';
-import { AlertIcon } from '../../svg/AlertIcon';
-import { SendImage } from '../../svg/SendImage';
 import { AuthContext } from '../../store/context';
 import { useReadStatus } from '../../hooks/useReadStatus';
-import { useAvatarArray } from '../../../src/hooks/useAvatarArray';
-import { Avatar } from '../../../src/components/Avatar/Avatar';
 // @ts-ignore
-import { EachChatMessage } from '@amityco/react-native-cli-chat-ui-kit/src/screens/ChatRoom/EachChatMessage';
+import { EachChatMessage } from './EachChatMessage';
+import { TopBar } from './TopBar';
+import { ChatRoomTextInput } from './ChatRoomTextInput';
 
 type ChatRoomScreenComponentType = React.FC<{}>;
 LogBox.ignoreLogs(['Warning: ...']); // Ignore log notification by message
@@ -128,8 +108,6 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
   const route = useRoute<RouteProp<RootStackParamList, 'ChatRoom'>>();
   const { chatReceiver, groupChat, channelId, from } = route.params;
 
-  const { avatarArray } = useAvatarArray(groupChat)
-
   const isGroupChat = useMemo(() => {
     return groupChat !== undefined;
   }, [groupChat]);
@@ -140,7 +118,6 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
   const [messagesData, setMessagesData] =
     useState<Amity.LiveCollection<Amity.Message>>();
   const [imageMultipleUri, setImageMultipleUri] = useState<string[]>([]);
-  const theme = useTheme() as MyMD3Theme;
 
   const {
     data: messagesArr = [],
@@ -197,9 +174,11 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
   const startRead = async () => {
     await SubChannelRepository.startReading(channelId);
   };
-  const stopRead = async () => {
+
+  const stopRead = useCallback(async () => {
     await SubChannelRepository.stopReading(channelId);
-  };
+  }, [channelId]);
+
   useEffect(() => {
     if (subChannelData && channelId) {
       startRead();
@@ -334,7 +313,7 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
     })()
   }, [messagesArr]);
 
-  const handleSend = async () => {
+  const handleSend = useCallback(async () => {
     setIsSendLoading(true);
     if (inputMessage.trim() === '') {
       return;
@@ -356,9 +335,9 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
       scrollToBottom();
       setIsSendLoading(false);
     }
-  };
+  }, [inputMessage, channelId]);
 
-  function handleBack(): void {
+  const handleBack = useCallback(() => {
     disposers.forEach((fn) => fn());
     stopRead();
     if (from === 'AddMembersFlow') {
@@ -373,7 +352,7 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
       //or else just go back
       navigation.goBack();
     }
-  }
+  }, [disposers, from, navigation])
 
   const loadNextMessages = () => {
     if (flatListRef.current && hasNextPage && onNextPage) {
@@ -394,17 +373,18 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
     setIsExpanded(!isExpanded);
   };
 
-  const scrollToBottom = () => {
+  const scrollToBottom = useCallback(() => {
     if (flatListRef && flatListRef.current) {
       (flatListRef.current as Record<string, any>).scrollToOffset({
         animated: true,
         offset: 0,
       });
     }
-  };
-  const handleOnFocus = () => {
+  }, [flatListRef]);
+
+  const handleOnFocus = useCallback(() => {
     setIsExpanded(false);
-  };
+  }, []);
 
   const pickCamera = async () => {
     const result: ImagePicker.ImagePickerResponse = await launchCamera({
@@ -528,60 +508,12 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
 
   return (
     <View style={styles.container}>
-      <View style={styles.topBar}>
-        <View style={styles.chatTitleWrap}>
-          <BackButton styles={styles.backButton} onPress={handleBack} />
-          {chatReceiver ? (
-            chatReceiver?.avatarFileId ? (
-              <Image
-                style={styles.avatar}
-                source={{
-                  uri: `https://api.${apiRegion}.amity.co/api/v3/files/${chatReceiver?.avatarFileId}/download`,
-                }}
-              />
-            ) : (
-              <View style={styles.avatar}>
-                <AvatarIcon />
-              </View>
-            )
-          ) : groupChat?.avatarFileId ? (
-            <Image
-              style={styles.avatar}
-              source={{
-                uri: `https://api.${apiRegion}.amity.co/api/v3/files/${groupChat?.avatarFileId}/download`,
-              }}
-            />
-          ) : (
-            <View style={styles.icon}>
-              <Avatar avatars={avatarArray} />
-            </View>
-          )}
-          <View>
-            <CustomText style={styles.chatName} numberOfLines={1}>
-              {chatReceiver
-                ? chatReceiver?.displayName
-                : groupChat?.displayName}
-            </CustomText>
-            {groupChat && (
-              <CustomText style={styles.chatMember}>
-                {groupChat?.memberCount} members
-              </CustomText>
-            )}
-          </View>
-        </View>
-        <TouchableOpacity
-          onPress={() => {
-            navigation.navigate('ChatDetail', {
-              channelId: channelId,
-              channelType: chatReceiver ? 'conversation' : 'community',
-              chatReceiver: chatReceiver ?? undefined,
-              groupChat: groupChat ?? undefined,
-            });
-          }}
-        >
-          <AlertIcon color={theme.colors.base} />
-        </TouchableOpacity>
-      </View>
+      <TopBar
+        chatReceiver={chatReceiver}
+        handleBack={handleBack}
+        groupChat={groupChat}
+        channelId={channelId}
+      />
       <View style={styles.chatContainer}>
         <FlatList
           data={sortedMessages}
@@ -605,61 +537,17 @@ const ChatRoom: ChatRoomScreenComponentType = () => {
           ListHeaderComponent={renderLoadingImages}
         />
       </View>
-      <KeyboardAvoidingView
-        behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
-        keyboardVerticalOffset={Platform.select({ ios: 50, android: 80 })}
-        style={styles.AllInputWrap}
-      >
-        <View style={styles.InputWrap}>
-          <TextInput
-            style={styles.input}
-            value={inputMessage}
-            onChangeText={(text) => setInputMessage(text)}
-            placeholder="Type a message..."
-            placeholderTextColor={theme.colors.baseShade3}
-            onFocus={handleOnFocus}
-          />
-
-          {inputMessage.length > 0 ? (
-            isSendLoading ? (
-              <ActivityIndicator style={styles.sendIcon} />
-            ) : (
-              <TouchableOpacity onPress={handleSend} style={styles.sendIcon}>
-                <SendChatIcon color={theme.colors.primary} />
-              </TouchableOpacity>
-            )
-          ) : (
-            <View>
-              <TouchableOpacity onPress={handlePress} style={styles.sendIcon}>
-                <SendImage color={theme.colors.base} />
-              </TouchableOpacity>
-            </View>
-          )}
-        </View>
-        {isExpanded && (
-          <View style={styles.expandedArea}>
-            <TouchableOpacity
-              onPress={pickCamera}
-              style={{ marginHorizontal: 30 }}
-            >
-              <View style={styles.IconCircle}>
-                <CameraBoldIcon color={theme.colors.base} />
-              </View>
-              <CustomText style={styles.iconText}>Camera</CustomText>
-            </TouchableOpacity>
-            <TouchableOpacity
-              // disabled={loadingImages.length > 0}
-              onPress={pickImage}
-              style={{ marginHorizontal: 20, alignItems: 'center' }}
-            >
-              <View style={styles.IconCircle}>
-                <AlbumIcon color={theme.colors.base} />
-              </View>
-              <CustomText style={styles.iconText}>Album</CustomText>
-            </TouchableOpacity>
-          </View>
-        )}
-      </KeyboardAvoidingView>
+      <ChatRoomTextInput
+        inputMessage={inputMessage}
+        isSendLoading={isSendLoading}
+        isExpanded={isExpanded}
+        setInputMessage={setInputMessage}
+        handleOnFocus={handleOnFocus}
+        handleSend={handleSend}
+        handlePress={handlePress}
+        pickCamera={pickCamera}
+        pickImage={pickImage}
+      />
       <ImageView
         images={[{ uri: fullImage }]}
         imageIndex={0}
